# erp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: ERP Platform API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: Token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: is_active
          schema:
            type: boolean
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          description: Forbidden
    post:
      tags: [Users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '403':
          description: Forbidden
        '409':
          description: Conflict
  /api/v1/users/{user_id}:
    put:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [Users]
      summary: Delete user (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/v1/users/{user_id}/reset-password:
    post:
      tags: [Users]
      summary: Reset user password
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserResetPasswordRequest'
      responses:
        '200':
          description: Reset user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/v1/warehouse/parts:
    get:
      tags: [Warehouse]
      summary: List parts with filters
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category_id
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated parts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Part'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
    post:
      tags: [Warehouse]
      summary: Create part
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartCreate'
      responses:
        '200':
          description: Created part
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
  /api/v1/warehouse/parts/import:
    post:
      tags: [Warehouse]
      summary: Import parts from CSV/XLSX (add-only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: mapping
          required: true
          schema:
            type: string
          description: JSON mapping of source columns to fields
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                upload:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
  /api/v1/maintenance/pm/generate-due:
    post:
      tags: [Maintenance]
      summary: Generate due PM work orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary of generated work orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateDueResponse'
  /api/v1/maintenance/work-orders/{work_order_id}:
    put:
      tags: [Maintenance]
      summary: Update work order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: work_order_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkOrderUpdate'
      responses:
        '200':
          description: Updated work order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
  /api/v1/tooling/batches/{batch_id}/operation:
    post:
      tags: [Tooling]
      summary: Execute batch operation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: batch_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchOperationPayload'
      responses:
        '200':
          description: Operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchOperationResult'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        password_change_required:
          type: boolean
      required: [access_token, refresh_token, token_type, password_change_required]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    ChangePasswordRequest:
      type: object
      properties:
        old_password:
          type: string
        new_password:
          type: string
      required: [old_password, new_password]
    UserRead:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        must_change_password:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, username, role, is_active, must_change_password, created_at, updated_at]
    UserRole:
      type: string
      enum: [user, admin, root]
    UserCreateRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        password:
          type: string
        must_change_password:
          type: boolean
      required: [username, role, password]
    UserUpdateRequest:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        password:
          type: string
        must_change_password:
          type: boolean
    UserResetPasswordRequest:
      type: object
      properties:
        temporary_password:
          type: string
        must_change_password:
          type: boolean
      required: [temporary_password]
    UserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserRead'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    Part:
      type: object
      properties:
        id:
          type: integer
        part_code:
          type: string
        name:
          type: string
        qty_on_hand:
          type: number
        min_stock:
          type: number
        price:
          type: number
        currency:
          type: string
    PartCreate:
      type: object
      properties:
        part_code:
          type: string
        name:
          type: string
        qty_on_hand:
          type: number
        min_stock:
          type: number
        price:
          type: number
        currency:
          type: string
      required: [part_code, name]
    ImportResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          type: integer
        errors:
          type: integer
    GenerateDueResponse:
      type: object
      properties:
        created_work_orders:
          type: integer
    WorkOrderUpdate:
      type: object
      properties:
        status:
          type: string
        summary:
          type: string
        downtime_min:
          type: number
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        summary:
          type: string
        downtime_min:
          type: number
    BatchOperationPayload:
      type: object
      properties:
        op_type:
          type: string
          enum: [inspection, cleaning, grinding]
        apply_to_all:
          type: boolean
        changes:
          type: array
          items:
            type: object
            properties:
              tool_id:
                type: integer
              changes:
                type: array
                items:
                  type: object
                  properties:
                    dim_name:
                      type: string
                    new_value:
                      type: number
      required: [op_type, changes]
    BatchOperationResult:
      type: object
      properties:
        processed:
          type: integer
        skipped:
          type: integer